separated evaluation from upload part
